+++ src/main/java/io/choerodon/agile/infra/dataobject/ExportIssuesDO.java
@@ -27,7 +27,7 @@ public class ExportIssuesDO {
     private String resolution;
     private BigDecimal sumEstimateTime;
     private BigDecimal sumRemainingTime;
-    private Integer storyPoints;
+    private BigDecimal storyPoints;
     private String epicName;
 
     public Long getIssueTypeId() {
@@ -190,11 +190,11 @@ public class ExportIssuesDO {
         this.epicName = epicName;
     }
 
-    public Integer getStoryPoints() {
+    public BigDecimal getStoryPoints() {
         return storyPoints;
     }
 
-    public void setStoryPoints(Integer storyPoints) {
+    public void setStoryPoints(BigDecimal storyPoints) {
         this.storyPoints = storyPoints;
     }
 }
